Don't mix with DLC's bossgroup
