+++ v1.4/Defs/BossGroupDefs/BossGroupsDef.xml
@@ -1,0 +1,55 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Defs>
+    
+    <RimArchive.BossgroupDef Name = "BA_BossGroupBase" Abstract = "True">
+        <quest>BA_Raid</quest>
+        <repeatWaveStartIndex>0</repeatWaveStartIndex>
+    </RimArchive.BossgroupDef>
+
+    <!-- HOD -->
+    <RimArchive.BossDef>
+        <defName>BA_HOD</defName>
+        <kindDef>BA_BOSS_HOD</kindDef>
+        <!-- appearAfterTicks默认为2^32-1 -->
+    </RimArchive.BossDef>
+
+    <RimArchive.BossgroupDef ParentName = "BA_BossGroupBase">
+        <defName>BA_BossGroup_HOD</defName>
+        <boss>BA_HOD</boss>
+        <waves> <!-- 下标从0开始 -->
+            <li>
+                <bossCount>1</bossCount>
+                <!-- 喽啰 -->
+                <escorts/>
+            </li>
+            <li>
+                <bossCount>1</bossCount>
+                <escorts>
+                    <Mech_Militor>5</Mech_Militor>
+                </escorts>
+                <bossApparel>
+                    <li>
+                        <thing>Apparel_HeavyShield</thing>
+                    </li>
+                </bossApparel>
+            </li>
+            <li>
+                <bossCount>1</bossCount>
+                <escorts>
+                    <Mech_Militor>7</Mech_Militor>
+                    <Mech_Pikeman>2</Mech_Pikeman>
+                </escorts>
+                <bossApparel>
+                    <li>
+                        <thing>Apparel_HeavyShield</thing>
+                    </li>
+                </bossApparel>
+            </li>
+        </waves>
+        <leaderDescription>The {LEADERKIND_label} can summon several pillars that applies debuffs to people around it.
+            Both {LEADERKIND_label} and its pillars are vulnerable to EMP and stun attacks. {LEADERKIND_label} will be stunned for a period of time, and the pillars will receive extra damage.
+        </leaderDescription>
+        <rewardDef>BA_Chip_HOD</rewardDef>
+    </RimArchive.BossgroupDef>
+
+</Defs>
