+++ v1.4/Defs/ThoughtDefs/Thought_Trait.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Defs>
+
+    <ThoughtDef>
+        <defName>BA_Hatred</defName>
+        <workerClass>ThoughtWorker_AlwaysActive</workerClass>
+        <validWhileDespawned>true</validWhileDespawned>
+        <developmentalStageFilter>Baby, Child, Adult</developmentalStageFilter>
+        <requiredTraits>
+            <li>BA_Trait_Arius</li>
+        </requiredTraits>
+        <stages>
+            <li>
+                <label>Vanity of Vanities</label>
+                <description>All is Vanity.</description>
+                <baseMoodEffect>-8</baseMoodEffect>
+            </li>
+        </stages>
+    </ThoughtDef>
+</Defs>
