+++ v1.4/Source/RimArchive/RimArchive/Boss/BossGroupDef.cs
@@ -1,0 +1,11 @@
+﻿using RimWorld;
+using Verse;
+
+namespace RimArchive;
+
+
+//又懒得自己写实现又不想和DLC的东西混一块。。。
+public class BossGroupDef : BossgroupDef
+{
+    new public BossDef boss;
+}
