+++ v1.4/Source/RimArchive/RimArchive/Defs/BossDef.cs
@@ -1,0 +1,9 @@
+﻿using Verse;
+using RimWorld;
+
+namespace RimArchive;
+
+public class BossDef : RimWorld.BossDef
+{
+
+}
