+++ v1.4/Source/RimArchive/RimArchive/GameComponents/RaidManager.cs
@@ -1,0 +1,9 @@
+﻿using RimWorld;
+using Verse;
+
+namespace RimArchive.GameComponents;
+
+public class RaidManager : GameComponent_Bossgroup
+{
+    public RaidManager(Game game) : base(game) { }
+}
